homework7

+++ src/news/comments/comments.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Param, Post, Get, Delete, Put, Res, UseInterceptors, UploadedFile, Render, ParseIntPipe } from '@nestjs/common';
+import { Body, Controller, Param, Post, Get, Delete, Put, Res, UseInterceptors, UploadedFile, Render, ParseIntPipe, UseGuards, Req, Query, HttpException, HttpStatus } from '@nestjs/common';
 import { CommentsService } from './comments.service';
 // import { Comment, CommentDto } from './comments.service'
 import { Response } from 'express'
@@ -9,6 +9,10 @@ import { diskStorage } from 'multer'
 import { HelperFileLoader } from '../../utils/HelperFileLoader';
 import { NewsService } from '../news.service';
 import { CommentsEntity } from './comments.entity';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { Roles } from 'src/auth/role/roles.decorator';
+import { Role } from 'src/auth/role/role.enum';
+
 
 const helperFileLoader = new HelperFileLoader();
 const PATH_NEWS = '/news_static';
@@ -18,6 +22,7 @@ helperFileLoader.path = PATH_NEWS;
 export class CommentsController {
     constructor(private readonly commentsService: CommentsService,
         private readonly newsService: NewsService,
+
     ) { }
 
     @Get('create/comment/:idNews/:idComm?')
@@ -43,55 +48,79 @@ export class CommentsController {
 
 
     @Post('/api/:idNews/:idComm?')
+    @UseGuards(JwtAuthGuard)
     async create(
         @Param('idNews', ParseIntPipe) idNews: number,
         @Body() comment: CreateCommentDto,
+        @Req() req,
         @Param('idComm') idComm?: string): Promise<CommentsEntity | Error> {
-        console.log('we are here. we are creating comment!')
-        console.log('commentcreate', comment);
         try {
             if (idComm) {
                 const idCommInt = parseInt(idComm);
             }
-            const newComment = await this.commentsService.create(idNews, comment);
+            const jwtUserId = await req.user.userId;
+            const newComment = await this.commentsService.create(idNews, comment.message, jwtUserId);
             return newComment
         } catch (error) {
             return new Error(`err: ${error}`);
         }
     }
 
-    @Get('api/:idNews')
-    // @Render('comment-list')
-    async get(@Param('idNews', ParseIntPipe) idNews: number) {
+    @Get('all')
+    async get(@Query('idNews', ParseIntPipe) idNews: number) {
         const news = await this.newsService.findById(idNews);
-        console.log('**** news ****', news)
         const comments = await this.commentsService.findAll(idNews)
-        console.log('comments', comments)
-        return { idNews: idNews, news, comments }
+        return comments
     }
 
     @Delete('api/:idComm')
-    async remove(@Param('idComm', ParseIntPipe) idComm: number): Promise<string | Error> {
+    @UseGuards(JwtAuthGuard)
+    @Roles(Role.User, Role.Admin)
+    async remove(
+        @Param('idComm', ParseIntPipe) idComm: number,
+        @Req() req): Promise<string | Error> {
         try {
-            const resultRemoving = await this.commentsService.remove(idComm)
-            return resultRemoving ? 'Комментарий удален!' : "Передан неверный идентификатор, удаление невозможно."
+
+            const jwtUserId = await req.user.userId;
+            const jwtUserRole = await req.user.userRoles;
+            const _comment = await this.commentsService.findById(idComm);
+
+            if (!_comment) {
+                throw new HttpException({
+                    status: HttpStatus.NOT_FOUND,
+                    error: 'Передан неверный идентификатор, удаление невозможно.'
+                },
+                    HttpStatus.NOT_FOUND
+                )
+            }
+
+            const commentIdUser = _comment.user.id;
+
+            if ((jwtUserRole === 'user' && commentIdUser === jwtUserId) || jwtUserRole === 'admin') {
+                const resultRemoving = await this.commentsService.remove(idComm)
+                return 'Комментарий удален!'
+            }
+            return 'Удаление невозможно'
         } catch (error) {
-            return new Error(`err: ${error}`);
+            throw new Error(`err: ${error}`);
         }
     }
 
+    @UseGuards(JwtAuthGuard)
     @Put('api/:idComm')
     async update(
         @Param('idComm', ParseIntPipe) idComm: number,
-        @Body() commentDto: EditCommentDto,
-        @Res() response: Response) {
+        @Body() message: EditCommentDto,
+        @Res() response: Response,
+        @Req() req) {
         try {
-            const comment = await this.commentsService.findById(idComm);
-            console.log(comment)
-            if (!comment) {
+            const _comment = await this.commentsService.findById(idComm);
+            const jwtUserId = await req.user.userId;
+            console.log('message', message.message, typeof message.message);
+            if (!_comment) {
                 response.status(404).end('Передан неверный идентификатор. Комментарий не найден')
             } else {
-                const result = await this.commentsService.update(idComm, commentDto);
+                const result = await this.commentsService.update(idComm, message.message);
                 if (result instanceof Error) {
                     response.status(500).end(result.message)
                 }
